==> polars_in_python/scripts_complete/03_begin_plotly.py <==
# %%
import polars as pl
import numpy as np
import plotly.express as px

# %%
# dat = pl.read_csv("../data/temple_details.csv", null_values="NA")\
#     .with_columns([
#         pl.col("prophet_start").str.slice(0,10)\
#             .str.to_datetime("%Y-%m-%d", utc=True)\
#             .alias("prophet_start"),
#         pl.col("prophet_end").str.slice(0,10)\
#             .str.to_datetime("%Y-%m-%d", utc=True)\
#             .alias("prophet_end"),
#         pl.col("announcement")\
#             .str.slice(0,4)\
#             .cast(pl.Int64())
#             .alias("announcement_year"),
#         pl.col("announcement")\
#             .str.to_datetime("%Y-%m-%d", utc=True)\
#             .alias("announcement"),
#         pl.col("groundbreaking")\
#             .str.to_datetime("%Y-%m-%d", utc=True)\
#             .alias("groundbreaking"),
#         pl.col("dedication")\
#             .str.to_datetime("%Y-%m-%d", utc=True)\
#             .alias("dedication"),
#     ])
# dat.write_parquet("../data/temple_details.parquet")
dat = pl.read_parquet("../data/temple_details.parquet")

# %%
df = px.data.iris() # iris is a pandas DataFrame
fig = px.scatter(df, x="sepal_width", y="sepal_length")
fig.write_html("../figures/figure.html")
fig.show()

# %%
df = px.data.gapminder().query("continent == 'Oceania'")
fig = px.line(df, x='year', y='lifeExp', color='country', markers=True)
fig.write_html("../figures/figure.html")
fig.show()

# %%
df = px.data.iris()
fig = px.scatter(df, x="sepal_length", y="sepal_width", marginal_x="histogram", marginal_y="rug")
fig.write_html("../figures/figure.html")
fig.show()

# %%
df = px.data.tips()
fig = px.scatter(df, x="total_bill", y="tip", color="smoker", facet_col="sex", facet_row="time")
fig.write_html("../figures/figure.html")
fig.show()

# %%
df = px.data.tips()
fig = px.scatter(df, x="total_bill", y="tip", trendline="ols")
fig.write_html("../figures/figure.html")
fig.show()

# %%
df = px.data.gapminder().query("year == 2007")
fig = px.treemap(df, path=[px.Constant("world"), 'continent', 'country'], values='pop',
                  color='lifeExp', hover_data=['iso_alpha'],
                  color_continuous_scale='RdBu',
                  color_continuous_midpoint=np.average(df['lifeExp'], weights=df['pop']))
fig.update_layout(margin = dict(t=50, l=25, r=25, b=25))
fig.write_html("../figures/figure.html")
fig.show()

# %%
df = px.data.tips()
fig = px.histogram(df, x="total_bill")
fig.write_html("../figures/figure.html")
fig.show()

# %%
df = px.data.tips()
fig = px.histogram(df, x="total_bill", color="sex", marginal="rug", # can be `box`, `violin`
                         hover_data=df.columns)
fig.write_html("../figures/figure.html")
fig.show()

# %%
df = px.data.tips()
fig = px.box(df, x="time", y="total_bill", points="all")
fig.write_html("../figures/figure.html")
fig.show()

# %%
df = px.data.tips()

fig = px.density_heatmap(df, x="total_bill", y="tip", marginal_x="histogram", marginal_y="histogram")
fig.write_html("../figures/figure.html")
fig.show()

# %%
df = px.data.iris()
fig = px.parallel_coordinates(df, color="species_id",
                              dimensions=['sepal_width', 'sepal_length', 'petal_width',
                                          'petal_length'],
                              color_continuous_scale=px.colors.diverging.Tealrose,
                              color_continuous_midpoint=2)
fig.write_html("../figures/figure.html")
fig.show()


# %%
df = px.data.tips()
fig = px.parallel_categories(df)
fig.write_html("../figures/figure.html")
fig.show()

# %%
# Styling
df = px.data.tips()
fig = px.histogram(df, x="day", y="total_bill", color="sex")
fig.write_html("../figures/figure.html")
fig.show()

# %%
df = px.data.tips()
fig = px.histogram(df, x="day", y="total_bill", color="sex",
            title="Receipts by Payer Gender and Day of Week",
            width=600, height=400,
            labels={ # replaces default labels by column name
                "sex": "Payer Gender",  "day": "Day of Week", "total_bill": "Receipts"
            },
            category_orders={ # replaces default order by column name
                "day": ["Thur", "Fri", "Sat", "Sun"], "sex": ["Male", "Female"]
            },
            color_discrete_map={ # replaces default color mapping by value
                "Male": "RebeccaPurple", "Female": "MediumPurple"
            },
            template="simple_white"
            )
fig.write_html("../figures/figure.html")
fig.show()

# %%
df = px.data.tips()
fig = px.histogram(df, x="day", y="total_bill", color="sex",
            title="Receipts by Payer Gender and Day of Week vs Target",
            width=600, height=400,
            labels={"sex": "Payer Gender",  "day": "Day of Week", "total_bill": "Receipts"},
            category_orders={"day": ["Thur", "Fri", "Sat", "Sun"], "sex": ["Male", "Female"]},
            color_discrete_map={"Male": "RebeccaPurple", "Female": "MediumPurple"},
            template="simple_white"
            )

fig.update_yaxes( # the y-axis is in dollars
    tickprefix="$", showgrid=True
)

fig.update_layout( # customize font and legend orientation & position
    font_family="Rockwell",
    legend=dict(
        title=None, orientation="h", y=1, yanchor="bottom", x=0.5, xanchor="center"
    )
)

fig.add_shape( # add a horizontal "target" line
    type="line", line_color="salmon", line_width=3, opacity=1, line_dash="dot",
    x0=0, x1=1, xref="paper", y0=950, y1=950, yref="y"
)

fig.add_annotation( # add a text callout with arrow
    text="below target!", x="Fri", y=400, arrowhead=1, showarrow=True
)
fig.write_html("../figures/figure.html")
fig.show()
